Fix give_feedback function
If applied, this commit fixes a bug in the feedback function which returned
an incorrect number of white pegs. The function now takes  into account the total
number of green pegs before computing the white peg count. It also does not display white
pegs for duplicate guesses if the corresponding duplicate is not found in the code.